+++ tapis-systemslib/src/main/java/edu/utexas/tacc/tapis/systems/service/SystemsService.java
@@ -103,6 +103,12 @@ public interface SystemsService
   Credential getUserCredential(ResourceRequestUser rUser, String systemId, String userName, AuthnMethod authnMethod)
           throws TapisException, TapisClientException, NotAuthorizedException;
 
+  String getGlobusAuthUrl(ResourceRequestUser rUser, String clientId)
+          throws TapisException, TapisClientException;
+
+  void generateAndSaveGlobusTokens(ResourceRequestUser rUser, String systemId, String userName, String authCode)
+          throws TapisException, TapisClientException;
+
   // ------------------- Scheduler Profiles---------------------------------
   // -----------------------------------------------------------------------
   void createSchedulerProfile(ResourceRequestUser rUser, SchedulerProfile schedulerProfile, String scrubbedText)